ref: Доработка и систематизация интерфейса пользователя.

+++ trunk/apps/frontend/modules/blog/templates/_comment.php
@@ -16,7 +16,14 @@
          || ($_blogContext->canDeleteSelf && ($_comment->web_user_id == $_blogContext->webUserId)))
         ? decorate_span('dangerAction', link_to('Удалить', $_blogContext->editorModule.'/deleteComment?id='.$_comment->id.'&returl='.$retUrlRaw, array('method' => 'delete', 'confirm' => 'Удалить комментарий?'))).' ' : '';
     echo Timing::dateToStr($_comment->create_time).' ';
-    echo decorate_span('safeAction', link_to_if(!$_blogContext->readOnly, $_comment->WebUser->login, 'webUser/show?id='.$_comment->web_user_id, array('target' => 'new')));
+    if ($_blogContext->readOnly)
+    {
+      echo decorate_span('info', $_comment->WebUser->login);
+    }
+    else
+    {
+      echo decorate_span('safeAction', link_to_if(!$_blogContext->readOnly, $_comment->WebUser->login, 'webUser/show?id='.$_comment->web_user_id, array('target' => 'new')));
+    }
     ?>
   </div>
